[UPG]-Upgrade module l10n_cr_base to Odoo v10
