--- l10n_cr_base/README.rst
@@ -1,31 +1,0 @@
-.. image:: https://img.shields.io/badge/licence-AGPL--3-blue.svg
-   :target: http://www.gnu.org/licenses/agpl-3.0-standalone.html
-   :alt: License: AGPL-3
-
-============
-l10n_cr_base
-============
-
-Features
---------
-
-* Adds a constraint for name and internal reference.
-* Sets internal reference fields as required for companies.
-
-Credits
-=======
-
-Contributors
-------------
-
-* Fabio Vilchez <fabio.vilchez@clearcorp.co.cr>
-
-
-Maintainer
-----------
-
-.. image:: https://avatars0.githubusercontent.com/u/7594691?v=3&s=200
-   :alt: ClearCorp
-   :target: http://clearcorp.cr
-
-This module is maintained by ClearCorp.
