+++ l10n_cr_base/__openerp__.py
@@ -1,23 +1,40 @@
 # -*- coding: utf-8 -*-
-# © <YEAR(S)> ClearCorp
-# License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl.html).
-
+##############################################################################
+#
+#    OpenERP, Open Source Management Solution
+#    Addons modules by CLEARCORP S.A.
+#    Copyright (C) 2009-TODAY CLEARCORP S.A. (<http://clearcorp.co.cr>).
+#
+#    This program is free software: you can redistribute it and/or modify
+#    it under the terms of the GNU Affero General Public License as
+#    published by the Free Software Foundation, either version 3 of the
+#    License, or (at your option) any later version.
+#
+#    This program is distributed in the hope that it will be useful,
+#    but WITHOUT ANY WARRANTY; without even the implied warranty of
+#    MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+#    GNU Affero General Public License for more details.
+#
+#    You should have received a copy of the GNU Affero General Public License
+#    along with this program.  If not, see <http://www.gnu.org/licenses/>.
+#
+##############################################################################
 {
     'name': 'l10n_cr_base',
-    'summary': 'Constraint for name and internal reference',
-    'version': '9.0.1.0',
-    'category': 'Base',
-    'website': 'http://clearcorp.cr',
+    'description': """ 
+l10n_cr_base
+=======================
+This module add a constraint for name and reference of partner.
+    """,
+    'version': '1.0',
     'author': 'ClearCorp',
-    'license': 'AGPL-3',
-    'sequence': 10,
-    'application': False,
+    'category': 'Base',
+    'website': "http://clearcorp.co.cr",
+    'depends': ['base'
+                ],
+    'data': ['res_partner.xml'
+             ],
+    'active': False,
     'installable': True,
-    'auto_install': False,
-    'depends': [
-        'base',
-    ],
-    'data': [
-        'views/res_partner.xml',
-    ],
+    'license': 'AGPL-3',
 }
