--- l10n_cr_base/models/__init__.py
@@ -1,5 +1,0 @@
-# -*- coding: utf-8 -*-
-# © 2016 ClearCorp
-# License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl.html).
-
-from . import res_partner
