--- l10n_cr_base/views/res_partner.xml
@@ -1,19 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<!-- © 2016 ClearCorp
-     License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl.html). -->
-
-<openerp>
-    <data>
-        <!--Form View Inherit res partner (base)-->
-        <record id="view_partner_form" model="ir.ui.view">
-            <field name="name">view.partner.form</field>
-            <field name="model">res.partner</field>
-            <field name="inherit_id" ref="base.view_partner_form"/>
-            <field name="arch" type="xml">
-                <field name="ref" position="attributes">
-                    <attribute name="attrs">{'required':[('is_company', '=', True),('parent_id', '=', False)]}</attribute>
-                </field>
-            </field>
-        </record>
-    </data>
-</openerp>
